move Engine base to utils.py file
